add tortilla category
Categories based on ethnicity does not really work for what I'm trying
to do. I don't have "mexican" as an ingredient in my pantry. I have
"tortilla", "beans", and "rice".

This isn't a substitute for a real, proper ingredient list application
thing, but it's close enough for a next-day "if I pick up this one
thing on the way home, I can make this".